Cookbook.Essential.Common

Description

global functions for the entirety of the Cookbook library. Common is the potpourri of Cookbook, with no category except everything uses me!

Documentation

sub :: Eq a => [a] -> Int -> [a]

Return a list starting at an index. Indices start at 0.

positions :: Eq a => [a] -> a -> [Int]

Find the occurrences of an element in a list.

pos :: Eq a => [a] -> a -> Int

C-style wrapper for positions. Returns the first occurrence in a list, or -1 on notElem.

apply :: [a -> a] -> a -> a

Reversal of map function. Chains calls of functions over a parameter, starting with head.

flt :: [[a]] -> [a]

Flatten a list one level.

fromLast :: ([a] -> [a]) -> [a] -> [a]

Execute a function from the end of a list to the front.

afterX :: Eq a => [a] -> a -> Int -> [a]

After a certain number of an element, return the list.

iSp :: [a] -> a -> [a]

Internal implementation of intersperse.