Cookbook.Essential.Continuous

Description

Library for overloading functions across lists and singular items, as well as tupples. Somewhat abuses FlexibleInstance and Typeclasses.

Documentation

class Continuous list part where

Classifies items that can be modified by either a list or item.

Methods

after :: list -> part -> list

Returns all elements after part.

before :: list -> part -> list

Returns all elements after part.

delete :: list -> part -> list

Removes part from the list.

class Splicable a b where

Classifies information which can be split by a tupple.

Methods

splice :: a -> b -> a

Removes everything between the tupple's parameters, including the parameters themselves.

class Replacable list repls where

Classifies data which can be replaced.

Methods

replace :: list -> repls -> list

Replaces part of list.

class Removable list toRm where

Classifies data which can be removed from a list.

Methods

remove :: list -> toRm -> list

Remove data from a list.