Cookbook.Essential.IO

Description

Library for completing common IO tasks, integrating with files and UNIX functions.

Documentation

filelines :: String -> IO [String]

Return the lines of a file as a list of Strings.

prompt :: String -> IO String

Prompts the user for keyboard input

inhome :: String -> IOMode -> IO Handle

Returns the path of a file in the user's home directory.

filename :: String -> String

Pure. Returns the file name with the directory truncated.

modulename :: String -> String

Pure. Returns the module name. That is, path to the file with the file cut off.