Cookbook.Ingredients.Functional.Break

Description

Library for breaking conditionally on lists. When one filters a list, it will filter all of the elements. When one filterBREAKS a list, it will stop collecting the list at a desired element. This library also includes functions for conditionally transforming a list.

Documentation

removeBreak :: (a -> Bool) -> [a] -> [a]

Drop a list until a predicate yields false, returning the false item and the rest of the list.

filterBreak :: (a -> Bool) -> [a] -> [a]

Collect a list until a predicate yields false for a value.

imbreak :: (a -> Bool) -> [a] -> Bool

Returns true if any element in the list yields true for a predicate.

btr :: (a -> Bool) -> (b, b) -> [a] -> [b]

Conditionally transform a list. If a predicate returns true, use lval. Otherwise, use rval.