Cookbook.Ingredients.Lists.Modify

Description

Library for modifying data within a list, and transforming lists in certain ways. While that's vague, so is the definition of "To Modify".

Documentation

rev :: [a] -> [a]

Reverses a list.

rm :: Eq a => [a] -> a -> [a]

Removes all occurances of an element from a list. See MDN1 in the source for a run-down on why it's implemented here and in Continuous.

splitOn :: Eq a => [a] -> a -> [[a]]

Splits a list on an element, making a list of lists based on the element as a seperator.

linesBetween :: Eq a => [[a]] -> ([a], [a]) -> [[a]]

Implementation of between that works on a list of lists, and using Contains rather than elem.

intersperse :: [a] -> a -> [a]

Put an element after every element of a list, not including the last element.

(?) :: a -> a -> Bool -> a

An equivelant to the terenary operator for Haskell.

surroundedBy :: Eq a => [a] -> (a, a) -> [[a]]

Returns a list of all elements surrounded by elements. Basically a between which works more than once.