Cookbook.Ingredients.Tupples.Assemble

Description

Library for arranging and modifying lists of twopples. Dubber Tuppers? There has to be a better word for this. It works on two-sided Tupples.

Documentation

tupsort :: Ord b => [(a, b)] -> [(a, b)]

Sorts a list of Tupples based on their second element.

assemble :: Ord b => [(a, b)] -> [a]

Order a list of tupples by their rval, and collect the lvals as a list.

rmDb :: Eq a => [(a, b)] -> [(a, b)]

Removes all double-entries from a list of tupples, contingent on just lval.