Cookbook.Ingredients.Tupples.Look

Description

Library for using tupples as an associative list. It can be thought of like a "map" in Clojure, or a database of key-value pairs. This is also the general-purpose tupple library, so there's stragglers.

Documentation

look :: Eq a => [(a, b)] -> a -> Maybe b

Look up an lval in a list of tupples.

lookList :: Eq a => [(a, b)] -> a -> [b]

Returns all elements where the lval matches the element.

swp :: (a, b) -> (b, a)

Swap the lval with the rval, and vice-versa.

rmLook :: Eq a => [(a, b)] -> a -> [(a, b)]

Removes all matches of the lval.

group :: [a] -> [(a, a)]

Turns a list into as many double-tupples (a,a) as it can, dropping items from uneven lists.