Cookbook.Project.Configuration.Configuration

Description

Configuration is a library for simple configuration files. It's pretty fragile, but it's been used for some pretty majorly used scripts, namely espion. The lvalue becomes the fst of a name-value pair, and the rvalue becomes the second, with the syntax: lvalue : rvaluen

Documentation

conf :: [String] -> String -> String

Read the lines of a configuration file, query it, and return an answer to the query.