Cookbook.Project.Quill2.Q2Api

Description

Q2Api is the user-facing part of Quill. It has all the functions necessar to Create, Read, Update, and Delete information from the database, and turn a database back into a Quill-readable string.

Documentation

getQuillName :: Quill -> String

Get the name of a Quill.

getQuillBody :: Quill -> Element String

Return the element of the Quill, specifically useful for lists.

getQuill :: [Quill] -> String -> QuillStatus Quill

Find a quill in the database by name, returning it or a possible error type.

lookUp :: [Quill] -> (String, String) -> QuillStatus String

Look up the value of a Quill TABLE. Will produce an error on a list.

removeQuill :: [Quill] -> String -> [Quill]

Remove a quill from the database by name.

removeItem :: [Quill] -> (String, String) -> QuillStatus [Quill]

Remove an item from a Quill within a database. Works aggressively, meaning it removes all copies to help sanitize QuillMultiples out. th

addQuill :: [Quill] -> Quill -> [Quill]

Adds a Quill databse to the file.

addItem :: [Quill] -> QuillAddition -> QuillStatus [Quill]

Add a QuillAddition to the databse. QuillAddition is a safe encapsulation of list and table values.

qMap :: QuillStatus [Quill] -> ([Quill] -> QuillStatus [Quill]) -> QuillStatus [Quill]

Map a Quill function.

changeItem :: [Quill] -> QuillAddition -> QuillStatus [Quill]

Change an item within the database using a Quill addition. Wrapper of addItem and removeItem.

toString :: Quill -> String

Turn a Quill table into a string.