Cookbook.Project.Quill2.Q2Io

Description

Q2Io is a helper library for interacting with Quill files. It aids in the reading of and writing to files on the user's system.

Documentation

fromFile :: FilePath -> IO [Quill]

Read a Quill database from a file.

toFile :: FilePath -> [Quill] -> IO ()

Send a Quill database into a parsable format in a file.