Cookbook.Project.Quill2.Q2Parse

Description

Q2Parse is the "engine" of Quill2. It is able to turn a string (or a list of them) into a Quill database. The syntax of Quill currently supports Whitespace-inclusive strings, comments, and whitespace-independence.

Documentation

decomment :: String -> String

Remove all C-style comments frm the code. Does not support single-line comments, because Quill2 is fully whitespace independant.

prepare :: [String] -> String

Prepare the lines of a file for processing.

pTable :: String -> Quill

Process a single entry in the database into a Table.

pFile :: [String] -> [Quill]

Turn the lines of a file into a list of tables, AKA a Database.