Cookbook.Project.Quill2.Q2Prelude

Description

Q2Prelude is the entry-level module for Quill2. It defines the data types that define data, databases, and errors throughout the rest of Quill2.

Documentation

data Element a

The body of a table or list.

Constructors

List [a]

Table [(a, a)]

type Quill = (String, Element String)

Helper type. Binds a name to a body.

data QuillStatus a

Encapsulates errors in the quill database. Currently supports Missing elements and Multiple Instances.

Constructors

QuillSuccess a

QuillMissing String

QuillMultiple String

data QuillAddition

Safe way of adding items to a Quill database. Allows type-checking on Lists and Tables when manipulating Elements.

Constructors

AList (String, String)

ATable (String, String, String)