Cookbook.Recipes.Algorithm

Description

Library for interacting with high-evel data structures found in Cookbook.Recipes.DataStructures (Ds). It also implements some data types that would have just been type synonyms if implemented in Ds.

Documentation

climb :: Tree a -> [a]

Get every node of a tree, put it into a list.

treeMap :: Tree a -> (a -> a) -> Tree a

Apply a function to every node in a tree.

treeFilter :: Tree a -> (a -> Bool) -> Tree a

Collectively nullify nodes on a tree.

genMatrix :: (Int, Int) -> a -> [((Int, Int), a)]

Generates a list of points, with the specified null data in the snd of the tupples.