Cookbook.Recipes.Cline

Documentation

parse :: String -> [Cline]

Parse a flat string into a list of Clines.

data Cline

Module : Cookbook.Recipes.Cline Copyright : (c) 2014 by Nate Pisarski License : BSD3 Maintainer : nathanpisarski@gmail.com Stability : Stable Portability : Portable (Cookbook) Cline is a library for managing strings meant to represent command-line arguments. Flags can toggle functionality in programs. Command-line arguments and their short names take an argument.

Represents command-line options. Works on either arguments or flags.

Constructors

Flag Char

Argument String String

clineExtract :: [String] -> [Cline]

Extract arguments from a list of word-separated strings.

clArg :: String -> String

Cleans a command line argument. If it's not a number, it will add a couple of dashes in front of it.