Cookbook.Recipes.Math

Description

Library for working with numbers more easily. It includes somewhat lazy functions for when a lambda will clutter code up too much, as well as more involved mathematical formulae.

Documentation

inc :: Enum a => a -> a

Increase value by one.

dec :: Enum a => a -> a

Decrease value by one,

sqr :: Num a => a -> a

Multiply a number by itself.

avg :: Fractional a => [a] -> a

Find the average of a group of Fractionals.

stdev :: (Fractional a, Floating a) => [a] -> a

Find the standard deviation of a list of data.

fact :: (Num a, Enum a) => a -> a

Factorial, from 1 to point.