Cookbook.Recipes.Sanitize

Description

Library for sanitizing data, mainly strings.

Documentation

blacklist :: Eq a => [a] -> [a] -> [a]

Restricts an entire list of information from appearing in another list bit-by-bit.

rmleading :: Eq a => [a] -> a -> [a]

Removes a leading character from a list.

up :: Eq a => ([a], [a]) -> [a] -> [a]

Refpos wrapper for two lists, last to first.

down :: Eq a => ([a], [a]) -> [a] -> [a]

Refpos wrapper for two lists, first to last.

rmdb :: Eq a => [a] -> [a]

Removes all doubles in the list, turning them into just one occurrence.

rmdbAll :: Eq a => [a] -> [a]

Wholly removes doubles from a list.

tolower :: String -> String

Moves a string to lower-case.

toupper :: String -> String

Moves a string to upper-case.

rmlws :: String -> String

Removes all of the leading whitespace from a string.

rmsymbols :: String -> String

Removes all symbols from a string.