Cookbook.Recipes.Cline

Documentation

parse :: String -> [Cline]

Parse a flat string into a list of Clines.

data Cline

Represents command-line options. Works on either arguments or flags.

Constructors

Flag Char

Argument String String

clineExtract :: [String] -> [Cline]

Extract arguments from a list of word-separated strings.

clArg :: String -> String

Cleans a command line argument. If it's not a number, it will add a couple of dashes in front of it.